+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - bug fix - tests files were ignored during bit add when they're weren't part of the files array and .gitignore contained a record with leading exclamation mark
 - improve handling of errors from compilers which return promises
+- symlink tester env in isolated envs 
 
 ## [0.12.8-dev.1] - 2018-03-05